@@ -2,7 +2,7 @@
     "name": "algorithms-exercises",
     "version": "1.0.0",
     "description": "The exercises for the Complete Intro to Algorithms",
-    "main": "index.html",
+    "main": "index.js",
     "scripts": {
         "start": "parcel index.html",
         "build": "parcel build index.html",
@@ -11,17 +11,17 @@
     },
     "dependencies": {
         "lodash": "4.17.21",
-        "parcel": "2.0.0-beta.2",
-        "react": "17.0.2",
-        "react-dom": "17.0.2",
-        "react-router-dom": "5.2.0",
+        "parcel": "2.8.3",
+        "react": "18.2.0",
+        "react-dom": "18.2.0",
+        "react-router-dom": "6.10.0",
         "xxhashjs": "0.2.2"
     },
     "devDependencies": {
-        "@babel/core": "7.12.0",
-        "jest": "^26.6.3"
+        "@babel/core": "7.21.4",
+        "jest": "^29.5.0"
     },
     "resolutions": {
-        "@babel/preset-env": "7.13.8"
+        "@babel/preset-env": "7.21.4"
     }
 }
